requirements.txt
<<<<<<< HEAD
 Flask>=2.2.0
Flask-Mail>=0.9.1
yt-dlp>=2023.12.1
requests>=2.28.0
python-dotenv>=1.0.0
Flask-Cors>=3.0.10
gunicorn>=20.1.0
=======
 Flask>=2.2.0
Flask-Mail>=0.9.1
yt-dlp>=2023.12.1
requests>=2.28.0
python-dotenv>=1.0.0
Flask-Cors>=3.0.10
gunicorn>=20.1.0
>>>>>>> 0e7f951cb12a91f0556f3cc9eccf6088adac7d55
